--productos cambios de tipo de cantidad

+++ Gestion_Administrativa_Api/Models/Productos.cs
@@ -27,7 +27,7 @@ public partial class Productos
 
     public decimal? TotalIva { get; set; }
 
-    public long? Cantidad { get; set; }
+    public decimal? Cantidad { get; set; }
 
     public virtual ICollection<DetallePrecioProductos> DetallePrecioProductos { get; set; } = new List<DetallePrecioProductos>();
